+++ src/main/java/pl/eizodev/app/services/TransactionService.java
@@ -2,8 +2,10 @@ package pl.eizodev.app.services;
 
 import pl.eizodev.app.entities.Transaction;
 
+import java.math.BigDecimal;
+
 public interface TransactionService {
     void addTransaction(Transaction transaction);
-    void performTransaction(String transactionType, String ticker, Long userId, float price, int quantity);
+    void performTransaction(String transactionType, String ticker, Long userId, BigDecimal price, int quantity);
     void deleteTransaction(Long id);
 }
