+++ src/main/java/pl/eizodev/app/services/TransactionServiceImpl.java
@@ -5,6 +5,7 @@ import pl.eizodev.app.entities.Transaction;
 import pl.eizodev.app.repositories.TransactionRepository;
 
 import javax.transaction.Transactional;
+import java.math.BigDecimal;
 import java.util.Optional;
 
 @Service
@@ -23,7 +24,7 @@ class TransactionServiceImpl implements TransactionService {
     }
 
     @Override
-    public void performTransaction(String transactionType, String ticker, Long userId, float price, int quantity) {
+    public void performTransaction(String transactionType, String ticker, Long userId, BigDecimal price, int quantity) {
         Optional<Transaction> existInDB = transactionRepository.findByTransactionTypeAndStockTicker(transactionType, ticker);
 
         if (existInDB.isPresent()) {
